+++ frontend/src/components/NewAdModal/modalComponents/AdModalForm.tsx
@@ -2,31 +2,23 @@ import {
   Box,
   Button,
   Divider,
-  FormControl,
-  FormHelperText,
-  FormLabel,
   IconButton,
-  InputLabel,
-  MenuItem,
-  Select,
-  Slider,
   Stack,
-  TextField,
   Typography,
   useMediaQuery,
 } from "@mui/material";
-import { useForm, Controller } from "react-hook-form";
-import {
-  AdInput,
-  useCreateAdMutation,
-  useGetAllSkillsQuery,
-} from "../../../generated/graphql-types";
-import type { AddressSuggestion, Skill } from "../../../types";
+import { useForm, FormProvider } from "react-hook-form";
+import { AdInput, useCreateAdMutation } from "../../../generated/graphql-types";
 import { useState } from "react";
 import theme from "../../../mui";
 import AdModalFormAddress from "./AdModalFormAddress";
 import { CameraAlt } from "@mui/icons-material";
 import CloseIcon from "@mui/icons-material/Close";
+import AdModalFormTitle from "./AdModalFormTitle";
+import AdModalFormDescription from "./AdModalFormDescription";
+import AdModalFormCategory from "./AdModalFormCategory";
+import AdModalFormDuration from "./AdModalFormDuration";
+import { AddressSuggestion } from "../../../types";
 
 interface AdModalFormProps {
   closeModal: () => void;
@@ -34,31 +26,11 @@ interface AdModalFormProps {
 
 export default function AdModalForm({ closeModal }: AdModalFormProps) {
   const isResponsiveLayout = useMediaQuery(theme.breakpoints.down("md"));
-  const {
-    control,
-    register,
-    handleSubmit,
-    watch,
-    setValue,
-    formState: { errors },
-  } = useForm<AdInput>({
+  const methods = useForm<AdInput>({
     defaultValues: { title: "", description: "", duration: 0 },
   });
 
-  const {
-    loading: skillsLoading,
-    error: skillsError,
-    data: skillsData,
-  } = useGetAllSkillsQuery();
   const [createAdMutation, { loading, error }] = useCreateAdMutation();
-
-  const skills: Skill[] = skillsData?.getAllSkills ?? [];
-
-  const [title, description] = watch("title", "description");
-
-  // Selection de l'adresse
-  const [selectedSuggestion, setSelectedSuggestion] =
-    useState<AddressSuggestion | null>(null);
 
   // Gestion des images et de leur prévisualisation
   const [files, setFiles] = useState<File[]>([]);
@@ -94,9 +66,12 @@ export default function AdModalForm({ closeModal }: AdModalFormProps) {
     setFileUrls(newFileUrls);
   };
 
+  // Selection de l'adresse
+  const [selectedSuggestion, setSelectedSuggestion] =
+    useState<AddressSuggestion | null>(null);
+
   // Gestion de la soumission du formulaire
   const onFormSubmitted = async (formData: AdInput) => {
-    console.log("files", setFiles);
     if (!selectedSuggestion) {
       console.error("Aucune adresse sélectionnée");
       return;
@@ -115,6 +90,9 @@ export default function AdModalForm({ closeModal }: AdModalFormProps) {
             longitude: selectedSuggestion.geometry.coordinates[0],
             duration: formData.duration,
             mangoAmount: formData.duration / 30,
+            // picture1: files[0] || null,
+            // picture2: files[1] || null,
+            // picture3: files[2] || null,
             skillId: formData.skillId,
             userRequesterId: "1", // TODO: get the user id from the context
           },
@@ -127,251 +105,146 @@ export default function AdModalForm({ closeModal }: AdModalFormProps) {
   };
 
   return (
-    <form onSubmit={handleSubmit(onFormSubmitted)}>
-      <Stack
-        direction={isResponsiveLayout ? "column" : "row"}
-        sx={{ margin: "3rem" }}
-      >
-        {/* Left column */}
+    <FormProvider {...methods}>
+      <form onSubmit={methods.handleSubmit(onFormSubmitted)}>
         <Stack
-          spacing={4}
-          sx={{
-            width: {
-              xs: "100%",
-              md: "60%",
-            },
-          }}
+          direction={isResponsiveLayout ? "column" : "row"}
+          sx={{ margin: "3rem" }}
         >
-          {/* Titre de l'annonce */}
-          <TextField
-            type="text"
-            label="Titre"
-            variant="standard"
-            {...register("title", {
-              required: "Champ obligatoire",
-              maxLength: {
-                value: 50,
-                message: "50 caractères maximum",
-              },
-            })}
-            onChange={(e) => {
-              const { value } = e.target;
-              setValue("title", value, { shouldValidate: true });
-            }}
-            placeholder="Titre de l'annonce"
-            error={!!errors.title || title?.length > 50}
-            helperText={errors.title?.message}
+          {/* Left column */}
+          <Stack
+            spacing={4}
             sx={{
-              backgroundColor: "var(--white)",
-            }}
-          />
-
-          {/* Description de l'annonce */}
-          <TextField
-            type="text"
-            label="Description"
-            variant="standard"
-            multiline
-            maxRows={3}
-            {...register("description", {
-              required: "Champ obligatoire",
-              maxLength: {
-                value: 255,
-                message: "255 caractères maximum",
+              width: {
+                xs: "100%",
+                md: "60%",
               },
-            })}
-            onChange={(e) => {
-              const { value } = e.target;
-              setValue("description", value, { shouldValidate: true });
             }}
-            placeholder="Description"
-            error={!!errors.description || description?.length > 255}
-            helperText={errors.description?.message}
-            sx={{ backgroundColor: "var(--white)" }}
-          />
-
-          {/* Adresse de l'annonce */}
-          <Controller
-            name="address"
-            control={control}
-            rules={{
-              required: "Champ obligatoire",
-            }}
-            render={({ field }) => (
-              <AdModalFormAddress
-                selectedSuggestion={selectedSuggestion}
-                onSuggestionChange={(value) => {
-                  field.onChange(value);
-                  setSelectedSuggestion(value);
-                }}
-                error={errors.address?.message}
-              />
-            )}
-          />
-
-          {/* Compétence requise pour l'annonce */}
-          <FormControl
-            error={!!errors.skillId}
-            variant="standard"
-            sx={{ backgroundColor: "var(--white)" }}
           >
-            <InputLabel>Catégorie</InputLabel>
-            <Select
-              {...register("skillId", { required: "Champ obligatoire" })}
-              defaultValue=""
-              disabled={skillsLoading || !!skillsError}
-            >
-              {skills.map((skill) => (
-                <MenuItem key={skill.id} value={skill.id}>
-                  {skill.name}
-                </MenuItem>
-              ))}
-            </Select>
-            <FormHelperText>{errors.skillId?.message}</FormHelperText>
-          </FormControl>
-        </Stack>
+            {/* Titre de l'annonce */}
+            <AdModalFormTitle />
 
-        {/* Divider */}
-        {!isResponsiveLayout && (
-          <Divider orientation="vertical" flexItem sx={{ margin: "0 4rem" }} />
-        )}
+            {/* Description de l'annonce */}
+            <AdModalFormDescription />
 
-        {/* Right column */}
-        <Stack
-          sx={{
-            width: {
-              xs: "100%",
-              md: "40%",
-            },
-          }}
-        >
-          {/* Durée de l'annonce */}
-          <Controller
-            name="duration"
-            control={control}
-            rules={{
-              required: "Veuillez sélectionner une durée.",
-              validate: (value) =>
-                value > 0 || "Veuillez sélectionner une durée.",
+            {/* Adresse de l'annonce */}
+            <AdModalFormAddress setSelectedSuggestion={setSelectedSuggestion} />
+
+            {/* Compétence requise pour l'annonce */}
+            <AdModalFormCategory />
+          </Stack>
+
+          {/* Divider */}
+          {!isResponsiveLayout && (
+            <Divider
+              orientation="vertical"
+              flexItem
+              sx={{ margin: "0 4rem" }}
+            />
+          )}
+
+          {/* Right column */}
+          <Stack
+            sx={{
+              width: {
+                xs: "100%",
+                md: "40%",
+              },
             }}
-            render={({ field }) => (
-              <FormControl>
-                <FormLabel sx={{ textAlign: "center" }}>DUREE</FormLabel>
-                <Slider
-                  {...field}
-                  size="small"
-                  sx={{ width: "80%", margin: "0 auto" }}
-                  valueLabelDisplay="auto"
-                  step={30}
-                  min={0}
-                  max={360}
-                  marks
-                  valueLabelFormat={(value) => {
-                    const hours = Math.floor(value / 60);
-                    const minutes = value % 60;
-                    if (hours === 0) {
-                      return `${minutes}min`;
-                    } else if (minutes === 0) {
-                      return `${hours}h`;
-                    } else {
-                      return `${hours}h ${minutes}min`;
-                    }
-                  }}
-                />
-              </FormControl>
-            )}
-          />
-          <FormHelperText>{errors.duration?.message}</FormHelperText>
+          >
+            {/* Durée de l'annonce */}
+            <AdModalFormDuration />
 
-          {/* Photos */}
-          <Typography color="rgba(0, 0, 0, 0.6)" sx={{ textAlign: "center" }}>
-            AJOUTER DES PHOTOS
-          </Typography>
-          <Stack direction="row" sx={{ justifyContent: "center" }}>
-            {/* Affichage des trois conteneurs pour les images */}
-            {[0, 1, 2].map((index) => (
-              <Box
-                key={index}
-                sx={{
-                  width: 92,
-                  height: 86,
-                  borderRadius: 2.5,
-                  backgroundColor: "#E0E0E0",
-                  border: "2px dotted #AFACAC",
-                  display: "flex",
-                  justifyContent: "center",
-                  alignItems: "center",
-                  position: "relative",
-                }}
-              >
-                {/* Si un fichier est sélectionné, afficher l'image */}
-                {files[index] ? (
-                  <>
-                    <img
-                      src={fileUrls[index]}
-                      alt={`photo-${index}`}
-                      style={{
-                        width: "100%",
-                        height: "100%",
-                        objectFit: "cover",
-                        borderRadius: 4,
-                      }}
-                    />
-                    {/* Icône de suppression */}
-                    <IconButton
-                      onClick={() => handleDelete(index)}
-                      sx={{
-                        position: "absolute",
-                        top: "0.1rem",
-                        right: "0.1rem",
-                        color: "white",
-                      }}
-                    >
-                      <CloseIcon />
-                    </IconButton>
-                  </>
-                ) : (
-                  <>
-                    {/* Si aucun fichier sélectionné, afficher l'icône d'appareil photo*/}
-                    <IconButton
-                      component="label"
-                      sx={{
-                        position: "absolute",
-                        zIndex: 1,
-                      }}
-                    >
-                      <CameraAlt
-                        sx={{
-                          width: 50,
-                          height: 50,
+            {/* Photos */}
+            <Typography color="rgba(0, 0, 0, 0.6)" sx={{ textAlign: "center" }}>
+              AJOUTER DES PHOTOS
+            </Typography>
+            <Stack direction="row" sx={{ justifyContent: "center" }}>
+              {/* Affichage des trois conteneurs pour les images */}
+              {[0, 1, 2].map((index) => (
+                <Box
+                  key={index}
+                  sx={{
+                    width: 92,
+                    height: 86,
+                    borderRadius: 2.5,
+                    backgroundColor: "#E0E0E0",
+                    border: "2px dotted #AFACAC",
+                    display: "flex",
+                    justifyContent: "center",
+                    alignItems: "center",
+                    position: "relative",
+                  }}
+                >
+                  {/* Si un fichier est sélectionné, afficher l'image */}
+                  {files[index] ? (
+                    <>
+                      <img
+                        src={fileUrls[index]}
+                        alt={`photo-${index}`}
+                        style={{
+                          width: "100%",
+                          height: "100%",
+                          objectFit: "cover",
+                          borderRadius: 4,
                         }}
                       />
-                      <input
-                        type="file"
-                        accept="image/*"
-                        style={{ display: "none" }}
-                        onChange={(e) => handleFileChange(e, index)}
-                      />
-                    </IconButton>
-                  </>
-                )}
-              </Box>
-            ))}
+                      {/* Icône de suppression */}
+                      <IconButton
+                        onClick={() => handleDelete(index)}
+                        sx={{
+                          position: "absolute",
+                          top: "0.1rem",
+                          right: "0.1rem",
+                          color: "white",
+                        }}
+                      >
+                        <CloseIcon />
+                      </IconButton>
+                    </>
+                  ) : (
+                    <>
+                      {/* Si aucun fichier sélectionné, afficher l'icône d'appareil photo*/}
+                      <IconButton
+                        component="label"
+                        sx={{
+                          position: "absolute",
+                          zIndex: 1,
+                        }}
+                      >
+                        <CameraAlt
+                          sx={{
+                            width: 50,
+                            height: 50,
+                          }}
+                        />
+                        <input
+                          type="file"
+                          accept="image/*"
+                          style={{ display: "none" }}
+                          onChange={(e) => handleFileChange(e, index)}
+                        />
+                      </IconButton>
+                    </>
+                  )}
+                </Box>
+              ))}
+            </Stack>
           </Stack>
         </Stack>
-      </Stack>
 
-      {/* Submit button */}
-      <Stack
-        sx={{
-          alignItems: "flex-end",
-          margin: "2.5rem",
-        }}
-      >
-        <Button type="submit">Envoyer</Button>
-      </Stack>
-      {loading && "Envoi en cours..."}
-      {error && "Une erreur est survenue, merci de réessayer..."}
-    </form>
+        {/* Submit button */}
+        <Stack
+          sx={{
+            alignItems: "flex-end",
+            margin: "2.5rem",
+          }}
+        >
+          <Button type="submit">Envoyer</Button>
+        </Stack>
+        {loading && "Envoi en cours..."}
+        {error && "Une erreur est survenue, merci de réessayer..."}
+      </form>
+    </FormProvider>
   );
 }
